% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr.R
\name{ngr.data.generator.bsplines}
\alias{ngr.data.generator.bsplines}
\title{Generating random curves from B-Splines}
\usage{
ngr.data.generator.bsplines(
  n,
  nknots,
  norder,
  p,
  domain = c(0, 1),
  snr,
  betaind
)
}
\arguments{
\item{n}{Number of curves}

\item{nknots}{Number of knots}

\item{norder}{Degree}

\item{p}{Number of time points}

\item{domain}{Domain of time}

\item{snr}{Signal to noise ratio}

\item{betaind}{Numeric index for function}
}
\value{
X: The generated X matrix of curve sampled at each timepoint

Y: The generated dependent variable
}
\description{
Generating random curves from B-Splines
n,nknots,norder,p,domain=c(0,1),snr,betaind
}
