% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truck}
\alias{truck}
\title{Truck emissions data}
\format{
A data frame with 108 rows and 91 columns:
\describe{
  \item{Y}{Emmission}
  \item{X1-X90}{Acceleration at each second}
}
}
\source{
Clark, N., Gautam, M., Wayne, W., Lyons, D., Thompson, G., and Zielinska,
B. (2007), “Heavy-Duty Vehicle Chassis Dynamometer Testing for Emissions Inventory, Air Quality Modeling, Source Apportionment and Air
Toxics Emissions Inventory: E55/59 All Phases,” Coordinating Research Council, Alpharetta
}
\usage{
truck
}
\description{
The particulate matter emissions data, taken from the Coordinating Research Councils
E55/E59 research project (Clark et al. 2007). In the project,
trucks were placed on the chassis dynamometer bed to mimic
inertia and particulate matter was measured by an emission analyzer on standard test cycles. The engine acceleration of diesel
trucks was also recorded.
}
\keyword{datasets}
